+++ src/main/java/com/airsaid/localization/config/SettingsState.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright 2021 Airsaid. https://github.com/airsaid
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 package com.airsaid.localization.config;
 
 import com.airsaid.localization.translate.AbstractTranslator;
@@ -46,6 +63,7 @@ public final class SettingsState implements PersistentStateComponent<SettingsSta
 
   public void initSetting() {
     TranslatorService translatorService = TranslatorService.getInstance();
+    translatorService.setUseGoogleComHost(isUseGoogleCom());
     AbstractTranslator selectedTranslator = translatorService.getSelectedTranslator();
     if (selectedTranslator == null) {
       LOG.info("initSetting");
@@ -62,6 +80,14 @@ public final class SettingsState implements PersistentStateComponent<SettingsSta
 
   public void setSelectedTranslator(AbstractTranslator translator) {
     this.state.selectedTranslatorKey = translator.getKey();
+  }
+
+  public boolean isUseGoogleCom() {
+    return state.isUseGoogleCom;
+  }
+
+  public void setUseGoogleCom(boolean useGoogleCom) {
+    state.isUseGoogleCom = useGoogleCom;
   }
 
   public void setAppId(@NotNull String translatorKey, @NotNull String appId) {
@@ -118,5 +144,6 @@ public final class SettingsState implements PersistentStateComponent<SettingsSta
     public Map<String, String> appIds = new HashMap<>();
     public boolean isEnableCache = true;
     public int maxCacheSize = 100;
+    public boolean isUseGoogleCom;
   }
 }