+++ src/main/java/com/airsaid/localization/translate/AbstractTranslator.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright 2021 Airsaid. https://github.com/airsaid
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 package com.airsaid.localization.translate;
 
 import com.airsaid.localization.config.SettingsState;
@@ -31,17 +48,22 @@ public abstract class AbstractTranslator implements Translator, TranslatorConfig
     }
 
     String requestUrl = getRequestUrl(fromLang, toLang, text);
-    String requestParams = getRequestParams(fromLang, toLang, text)
-        .stream()
-        .map(pair -> pair.first.concat("=").concat(pair.second))
-        .collect(Collectors.joining("&"));
-
     RequestBuilder requestBuilder = HttpRequests.post(requestUrl, CONTENT_TYPE);
     configureRequestBuilder(requestBuilder);
 
     try {
       return requestBuilder.connect(request -> {
-        request.write(requestParams);
+        String requestParams = getRequestParams(fromLang, toLang, text)
+            .stream()
+            .map(pair -> pair.first.concat("=").concat(pair.second))
+            .collect(Collectors.joining("&"));
+        if (!requestParams.isEmpty()) {
+          request.write(requestParams);
+        }
+        String requestBody = getRequestBody(fromLang, toLang, text);
+        if (!requestBody.isEmpty()) {
+          request.write(requestBody);
+        }
         String resultText = request.readString();
         return parsingResult(fromLang, toLang, text, resultText);
       });
@@ -57,8 +79,18 @@ public abstract class AbstractTranslator implements Translator, TranslatorConfig
   }
 
   @Override
+  public boolean isNeedAppId() {
+    return true;
+  }
+
+  @Override
   public @Nullable String getAppId() {
     return SettingsState.getInstance().getAppId(getKey());
+  }
+
+  @Override
+  public boolean isNeedAppKey() {
+    return true;
   }
 
   @Override
@@ -76,6 +108,11 @@ public abstract class AbstractTranslator implements Translator, TranslatorConfig
 
   @NotNull
   public abstract List<Pair<String, String>> getRequestParams(@NotNull Lang fromLang, @NotNull Lang toLang, @NotNull String text);
+
+  @NotNull
+  public String getRequestBody(@NotNull Lang fromLang, @NotNull Lang toLang, @NotNull String text) {
+    return "";
+  }
 
   public abstract void configureRequestBuilder(@NotNull RequestBuilder requestBuilder);