+++ src/main/java/com/airsaid/localization/translate/services/TranslatorService.java
@@ -1,8 +1,26 @@
+/*
+ * Copyright 2021 Airsaid. https://github.com/airsaid
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 package com.airsaid.localization.translate.services;
 
 import com.airsaid.localization.translate.AbstractTranslator;
 import com.airsaid.localization.translate.impl.baidu.BaiduTranslator;
 import com.airsaid.localization.translate.impl.google.GoogleTranslator;
+import com.airsaid.localization.translate.impl.microsoft.MicrosoftTranslator;
 import com.airsaid.localization.translate.impl.youdao.YoudaoTranslator;
 import com.airsaid.localization.translate.lang.Lang;
 import com.intellij.openapi.application.ApplicationManager;
@@ -12,7 +30,7 @@ import com.intellij.openapi.diagnostic.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.function.Consumer;
 
@@ -31,11 +49,14 @@ public final class TranslatorService {
   private boolean isEnableCache = true;
 
   public TranslatorService() {
-    translators = new HashMap<>();
+    translators = new LinkedHashMap<>();
 
     GoogleTranslator googleTranslator = new GoogleTranslator();
     translators.put(googleTranslator.getKey(), googleTranslator);
     defaultTranslator = googleTranslator;
+
+    MicrosoftTranslator microsoftTranslator = new MicrosoftTranslator();
+    translators.put(microsoftTranslator.getKey(), microsoftTranslator);
 
     BaiduTranslator baiduTranslator = new BaiduTranslator();
     translators.put(baiduTranslator.getKey(), baiduTranslator);
@@ -94,6 +115,10 @@ public final class TranslatorService {
     LOG.info(String.format("doTranslate result: %s", result));
     cacheService.put(getCacheKey(fromLang, toLang, text), result);
     return result;
+  }
+
+  public void setUseGoogleComHost(boolean isUseGoogleComHost) {
+    GoogleTranslator.setUseComHost(isUseGoogleComHost);
   }
 
   public void setEnableCache(boolean isEnableCache) {